+++ components/LazyMint.tsx
@@ -2,8 +2,6 @@ import { contract_address, contract_abi } from "config";
 import { ImportCandidate } from "ipfs-core-types/src/utils";
 import { create } from "ipfs-http-client";
 import Web3 from "web3";
-// Context
-import { useCurrency } from "@/context/currencyContext";
 
 // @ts-expect-error
 const client = create("https://ipfs.infura.io:5001/api/v0");
@@ -13,15 +11,12 @@ export async function getVoucher(
   uri: string,
   minPrice = 0,
   signer: any,
-  currency: any,
+  currency: any
 ) {
-  console.log(currency)
-  // const { currency } = useCurrency();
   const domain = {
     name: "LazyNFT-Voucher",
     version: "0.01",
     verifyingContract: contract_address,
-    // chainId: parseInt(currency.chain_id),
   };
 
   const voucher = { tokenId, uri, minPrice };
@@ -47,49 +42,13 @@ export async function uploadIPFS(file: ImportCandidate) {
   return url;
 }
 
-// export async function availableToWithdraw() {
-//   const { currency } = useCurrency();
-//   const web3 = new Web3(currency.rpc_url);
-//   const contract = new web3.eth.Contract(
-//     // @ts-expect-error
-//     contract_abi,
-//     currency.contract_address
-//   );
-//   let balance = await contract.methods
-//     .availableToWithdraw()
-//     .call({ from: window.ethereum.selectedAddress });
-//   console.log(balance);
-// }
-
-// export async function withdrawTokens() {
-//   const { currency } = useCurrency();
-//   const web3 = new Web3(currency.rpc_url);
-//   const contract = new web3.eth.Contract(
-//     // @ts-expect-error
-//     contract_abi,
-//     currency.contract_address
-//   );
-//   contract.methods.withdraw();
-//   const params = {
-//     from: window.ethereum.selectedAddress,
-//     to: contract.options.address,
-//     chainId: currency.chain_id,
-//   };
-//   // contract.methods.withdraw();
-// }
-
 export async function redeemNFT(voucher, currency) {
-  console.log("Reached into function");
-  // console.log(currency);
-  console.log(currency.contract_address);
-  // const { currency } = useCurrency();
   const web3 = new Web3(currency.rpc_url);
   const contract = new web3.eth.Contract(
     // @ts-expect-error
     contract_abi,
     currency.contract_address
   );
-  console.log("Created new Web3 instance");
   const amount = voucher["minPrice"];
   const amountToSend = web3.utils.toWei(amount.toString(), "ether"); // Convert to wei value
   const params = {
@@ -99,15 +58,10 @@ export async function redeemNFT(voucher, currency) {
     data: contract.methods
       .redeem(window.ethereum.selectedAddress, voucher)
       .encodeABI(),
-    // chainId: currency.chain_id,
-    // gas: "2100000000",
   };
   let gasAmount = await web3.eth.estimateGas(params);
-  console.log("Gas Predicted");
   const res = await window.ethereum.request({
     method: "eth_sendTransaction",
-    params: [params,{gas:gasAmount}],
+    params: [params, { gas: gasAmount }],
   });
-  console.log("Done");
-  console.log(res);
 }
